Implementing a struct on MissingTCA
